no commit message

+++ CCFlow/WF/Comm/RefFunc/EnOnly.js
@@ -5,13 +5,13 @@ function ShowMultiFile(AttrFiles, FileManagers, isExitMyNum) {
     //如果多附件没有分类
     if (AttrFiles.length == 0) {
         html += "<div class='layui-row FoolFrmFieldRow'>";
-        html += "<div class='layui-col-md2 layui-col-xs2 FoolFrmFieldLabel'><label class='layui-form-label'>多附件上传</label></div>";
-       
+        html += "<div class='layui-col-md2 layui-col-xs2 FoolFrmFieldLabel'><label class='layui-form-label'>附件</label></div>";
+
         html += "<div class='layui-col-md10 layui-col-xs10 FoolFrmFieldInput'>";
         if (isExitMyNum == true) {
             html += "<div class='FoolFrmFieldLabel'>";
-            
-            if (mapData.GetPara("IsDelete") == 1 || mapData.GetPara("IsUpdate") == 1 || mapData.GetPara("IsInsert") == 1)
+
+            if (isReadonly != "1" && (mapData.GetPara("IsDelete") == 1 || mapData.GetPara("IsUpdate") == 1 || mapData.GetPara("IsInsert") == 1))
                 html += "<a href='javaScript:void(0)' style='color:#fff;' class='layui-btn layui-btn-xs' onclick='ActiveUploadFile(\"\",\"\")'>上传附件</a>";
 
             html += "</div>"
@@ -24,9 +24,10 @@ function ShowMultiFile(AttrFiles, FileManagers, isExitMyNum) {
             html += "<td width='*'>";
             html += "<img src='../../Img/FileType/" + sf.MyFileExt.substr(1) + ".gif' border=0 />" + sf.MyFileName + sf.MyFileExt;
             html += "</td>";
+            html += "<td width='25%'>" + sf.RDT + "</td>";
             html += "<td width='25%'>";
             html += "<a href='javaScript:void(0)' onclick='downLoadFileM(\"" + sf.OID + "\")'>下载</a>&nbsp;&nbsp;&nbsp;&nbsp;";
-            if (mapData.GetPara("IsDelete") == 1 || mapData.GetPara("IsUpdate") == 1 || mapData.GetPara("IsInsert") == 1 )
+            if (isReadonly != "1" && (mapData.GetPara("IsDelete") == 1 || mapData.GetPara("IsUpdate") == 1 || mapData.GetPara("IsInsert") == 1))
                 html += "<a href='javaScript:void(0)' onclick='deleteFile(\"" + sf.OID + "\",this)'>删除</a>";
             html += "</td>";
             html += "</tr>";
@@ -36,7 +37,7 @@ function ShowMultiFile(AttrFiles, FileManagers, isExitMyNum) {
         if (count == 0)
             html += "没有上传附件";
         html += "</div>";
-        
+
         html += "</div>";
         return html;
     }
@@ -44,12 +45,12 @@ function ShowMultiFile(AttrFiles, FileManagers, isExitMyNum) {
     for (var i = 0; i < AttrFiles.length; i++) {
         var attrFile = AttrFiles[i];
         html += "<div class='layui-row FoolFrmFieldRow'>";
-        html += "<div class='layui-col-md2 layui-col-xs2 FoolFrmFieldLabel'><label class='layui-form-label'>多附件上传</label></div>";
+        html += "<div class='layui-col-md2 layui-col-xs2 FoolFrmFieldLabel'><label class='layui-form-label'>附件</label></div>";
 
         html += "<div class='layui-col-md10 layui-col-xs10 FoolFrmFieldInput'>";
         if (isExitMyNum == true) {
             html += "<div class='FoolFrmFieldLabel'>";
-            if (mapData.GetPara("IsDelete") == 1 || mapData.GetPara("IsUpdate") == 1 || mapData.GetPara("IsInsert") == 1)
+            if (isReadonly != "1" && (mapData.GetPara("IsDelete") == 1 || mapData.GetPara("IsUpdate") == 1 || mapData.GetPara("IsInsert") == 1))
                 html += "<a href='javaScript:void(0)' style='color:#fff;' class='layui-btn layui-btn-xs' onclick='ActiveUploadFile(\"" + attrFile.FileNo + "\",\"" + attrFile.FileName + "\")'>上传附件</a>";
             html += "</div>"
         }
@@ -91,9 +92,10 @@ function ActiveUploadFile(upfileNo, upfileName) {
         if (flag == false)
             return;
     }
-       
+
     fileNo = upfileNo;
     fileName = upfileName;
+    $("#File_Upload").val("");
     //激活上传文件
     $("#File_Upload").click();
 }
@@ -102,10 +104,7 @@ function downLoadFileM(OID) {
     if (plant == "CCFlow")
         SetHref('../../CCForm/DownFile.aspx?DoType=EntityMutliFile_Load&OID=' + OID);
     else {
-        var currentPath = GetHrefUrl();
-        var path = currentPath.substring(0, currentPath.indexOf('/WF') + 1);
-        Url = path + '/WF/CCForm/EntityMutliFile_Load.do?OID=' + OID;
-        SetHref(url);
+        SetHref(basePath + '/WF/Ath/EntityMutliFile_Load.do?OID=' + OID);
     }
 }
 
@@ -123,6 +122,7 @@ function deleteFile(OID, td) {
     }
     //处理显示问题，删除一行
     $(td).parent().parent().remove();
+    window.location.reload();
 
 }
 
@@ -138,9 +138,10 @@ function tableAddTr(fileManager) {
     html += "<td>";
     html += "<img src='../../Img/FileType/" + fileManager.MyFileExt.substr(1) + ".gif' border=0 />" + fileManager.MyFileName + fileManager.MyFileExt;
     html += "</td>";
+    html += "<td>" + fileManager.RDT + "</td>";
     html += "<td>";
     html += "<a href='javaScript:void(0)' onclick='downLoadFileM(\"" + fileManager.OID + "\")'>下载</a>&nbsp;&nbsp;&nbsp;&nbsp;";
-    if (mapData.GetPara("IsDelete") == 1)
+    if (mapData.GetPara("IsDelete") == 1 || mapData.GetPara("IsUpdate") == 1 || mapData.GetPara("IsInsert") == 1)
         html += "<a href='javaScript:void(0)' onclick='deleteFile(\"" + fileManager.OID + "\",this)'>删除</a>";
     html += "</td>";
     html += "</tr>";
@@ -151,70 +152,22 @@ function tableAddTr(fileManager) {
 //多文件上传
 function MultiUploadFile() {
 
-    var fileObj = document.getElementById("File_Upload").files[0]; // js 获取文件对象
-    if (fileObj == null || (fileObj) == "undefined" || fileObj.size <= 0) {
+    var handler = new HttpHandler("BP.WF.HttpHandler.WF_CommEntity");
+    handler.AddFileData();
+    handler.AddPara("FileNo", fileNo);
+    handler.AddPara("FileName", fileName);
+    handler.AddPara("EnName", GetQueryString("EnName"));
+    handler.AddPara("PKVal", pkVal);
+    var data = handler.DoMethodReturnString("EntityMultiAth_Upload");
+    if (data.indexOf("err@") != -1) {
+        alert(data);
         return;
     }
-
-    //获取文件的后缀名
-    if (fileObj.name.indexOf(".") == -1) {
-        alert("上传的文件名没有扩展名，请检查后上传");
-        return;
-    }
-
-
-    //form表单序列话
-    var formData = new FormData();
-    var name = $("input").val();
-    formData.append("file", fileObj);
-    formData.append("name", name);
-    //获取发送请求时的参数
-    var queryString = document.location.search.substr(1);
-
-    queryString = replaceParamVal(queryString, "PKVal", pkVal);
-
-    queryString += "&FileNo=" + fileNo + "&FileName=" + encodeURI(fileName);
-
-    //URL 路径
-    var URL = basePath + "/WF/Comm/Handler.ashx?DoType=EntityMultiAth_Upload";
-    if (plant != "CCFlow") {
-        var currentPath = GetHrefUrl();
-        var path = currentPath.substring(0, currentPath.indexOf('/WF') + 1);
-        URL = path + "/WF/Comm/ProcessRequest.do?DoType=EntityMultiAth_Upload";
-    }
-
-    if (queryString != null && queryString.length > 0)
-        URL += "&" + queryString;
-
-
-    $.ajax({
-        url: URL,
-        type: 'POST',
-        data: formData,
-        async: false,
-        xhrFields: {
-            withCredentials: true
-        },
-        crossDomain: true,
-        // 告诉jQuery不要去处理发送的数据
-        processData: false,
-        // 告诉jQuery不要去设置Content-Type请求头
-        contentType: false,
-        beforeSend: function () {
-            console.log("正在进行，请稍候");
-        },
-        success: function (responseStr) {
-            if (responseStr.indexOf('err@') == 0)
-                return;
-            responseStr = JSON.parse(responseStr);
-            tableAddTr(responseStr);
-        },
-        error: function (responseStr) {
-            console.log("error");
-        }
-    });
+    data = JSON.parse(data);
+    tableAddTr(data);
 
 }
+
 
 //树形结构
 function findChildren(jsonArray, parentNo) {
